[BEAM-1871] Hide internal implementation details of how we create a DefaultBucket for GCP Temp Location
Moved relevant contents of GcpProjectUtil and DefaultProject into GcpOptions.GcpTempLocation
